requirements.txt
fastapi
uvicorn[standard]
pandas
openpyxl
selenium
webdriver-manager
python-dotenv
requests
rapidfuzz